// src/common/enums/NeonAlertPlacement.ts
/**
 * Placement values for use with the <a href="/feedback/alert#examples">NeonAlert</a> component. This is used to
 * describe the positioning of alerts on the screen.
 */
export enum NeonAlertPlacement {
  /** Position alerts in the <strong>top left</strong> of the screen */
  TopLeft = 'top-left',
  /** Position alerts in the <strong>top right</strong> of the screen */
  TopRight = 'top-right',
  /** Position alerts in the <strong>bottom left</strong> of the screen */
  BottomLeft = 'bottom-left',
  /** Position alerts in the <strong>bottom right</strong> of the screen */
  BottomRight = 'bottom-right',
}